requirements.txt
solana
base58
pynacl